backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
anthropic==0.94.0
snowflake-connector-python==4.4.0
python-dotenv==1.2.2
pydantic==2.12.5
requests==2.33.1
ShopifyAPI==12.7.0
python-multipart==0.0.20
